Fix negative deficit problem

+++ packages/remotemetrics/events.go
@@ -94,11 +94,11 @@ type MessageScheduledMetrics struct {
 	// ReceivedTimestamp - IssuedTimestamp in nanoseconds
 	DeltaReceivedIssued int64 `json:"DeltaReceivedIssued" bson:"DeltaReceivedIssued"`
 	// ScheduledTimestamp - QueuedTimestamp in nanoseconds
-	SchedulingTime  int64   `json:"schedulingTime" bson:"schedulingTime"`
-	AccessMana      float64 `json:"accessMana" bson:"accessMana"`
-	StrongEdgeCount int     `json:"strongEdgeCount" bson:"strongEdgeCount"`
-	WeakEdgeCount   int     `json:"weakEdgeCount,omitempty" bson:"weakEdgeCount"`
-	LikeEdgeCount   int     `json:"likeEdgeCount,omitempty" bson:"likeEdgeCount"`
+	SchedulingTime  int64 `json:"schedulingTime" bson:"schedulingTime"`
+	AccessMana      int64 `json:"accessMana" bson:"accessMana"`
+	StrongEdgeCount int   `json:"strongEdgeCount" bson:"strongEdgeCount"`
+	WeakEdgeCount   int   `json:"weakEdgeCount,omitempty" bson:"weakEdgeCount"`
+	LikeEdgeCount   int   `json:"likeEdgeCount,omitempty" bson:"likeEdgeCount"`
 }
 
 // MissingMessageMetrics defines message solidification record that is sent to the remote logger.
